Create Translation Mechanism.py
This is a Python program. Often, especially in multilingual nations like India, text may be in the Roman script but not in the English language. To convert this to English, I used the Google Translate API (which I purchased) to detect the language of the text. If it wasn't English, then I transliterated the text into the detected language and then used the Google Translate API again to translate it into English. As far as I could tell, no direct conversion method existed for an Indian language in the Roman script to English (at the very least not for Python). So, I used two different kinds of tools. This program is only for Hindi, but can very easily be extended to any set of languages needed.