Create NLP Analysis With Google
This file encompasses two of the most crucial aspects of sentiment analysis, especially aspect based sentiment analysis. Firstly, it uses the Google Natural Language API to extract the entities and their associated sentiment scores from the text. It then deals with the extraction of each individual token, its part of speech (POS) and the root form of the word (lemma). It also shows how to chunk noun phrases (like converting 'billing' and 'counter' to 'billing counter').