Create SentiWordNet finder
Often, the need arises to find the sentiment of one specific word rather than entities within a sentence. This can be done through a resource that the Natural Language Tool Kit (NLTK) library gives access to in Python - SentiWordNet (SWN). 

SWN is a database of thousands of word lemmas that have a positive, negative and objective score associated with them. These can be understood as the probabilities that a word conveys positive, negative or neutral sentiments. This code is designed specifically to check whether a word conveys any sentiment at all.

It needs a POS and lemma for the word. Though we used the ones provided by Google, NLTK has functions for the same (PorterStemmer and Lemmatizer). SWN has its own way of tagging POS, and so conversion mechanisms are needed. Two of these are given here - from the Google system and from the Penn Treebank system.